Update ttt.py to run on windows
To run this example game on a windows machine, first install the curses library with 'pip install windows-curses'. Then run the program with 'python' or 'py', it will not run with 'python3'. 
ttt.py has been updated to handle unexpected exceptions thrown by windows when attempting to receive from a non-blocking socket that has not received anything.