Attempt fixing regression with on non-Windows
Sigh; it seems that WindowsError isn't defined on non-Windows, so you
can't try to catch it and just expect that it'll never be caught.
Fortunately, it turns out that Python translates errno values from Windows
versions to POSIXy equivalents.  The gotcha was that on many Unix-like
systems, EAGAIN and EWOULDBLOCK are the same; but on Windows they aren't.